pkg/cli: improve concurrency of tabledump upload
Currently, there the table dumps are read concurrency. One go routine
reads one table dumps at a time and uploads the parsed content in
batches. These batches are uploaded serially. This can get very slow
when the size of the table is large.

So, this commit fixes that by introducing two go routine pools. 1 is
responsible for reading the table dumps and parsing them. The workers
then hand off the parsed content in batches to the 2nd pool that uploads
them to datadog. This way, the upload of batches also happen
concurrently.

Epic: none
Release note: None